% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{add_blq_layers}
\alias{add_blq_layers}
\title{Add LOQ reference line and caption to a plot}
\usage{
add_blq_layers(
  plot,
  caption,
  loq_method,
  loq,
  dosenorm,
  loq_el,
  show_legend = FALSE
)
}
\arguments{
\item{plot}{ggplot object}

\item{caption}{character, current caption string}

\item{loq_method}{numeric, 0/1/2}

\item{loq}{numeric, lower limit of quantification value}

\item{dosenorm}{logical, whether dose normalization is active}

\item{loq_el}{A \code{pmx_line} element with LOQ line aesthetics.}

\item{show_legend}{logical, whether to add LLOQ to the linetype legend}
}
\value{
A named list with elements \code{plot} (modified ggplot) and \code{caption} (modified string)
}
\description{
Conditionally adds an horizontal reference line at the LLOQ and appends
BLQ imputation method text to the plot caption.
}
\keyword{internal}
