% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{add_obs_point_layer}
\alias{add_obs_point_layer}
\title{Internal helper: Add observation point layer to a DV-vs-concentration plot}
\usage{
add_obs_point_layer(plot, point_el, col_var_str)
}
\arguments{
\item{plot}{ggplot object to modify}

\item{point_el}{A \code{pmx_point} element with point aesthetics.}

\item{col_var_str}{String name of the color variable, or \code{NULL}}
}
\value{
The modified ggplot object
}
\description{
Internal helper: Add observation point layer to a DV-vs-concentration plot
}
\keyword{internal}
