% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvconc.R
\name{caption_dvconc}
\alias{caption_dvconc}
\title{Define a caption for \code{plot_dvconc}}
\usage{
caption_dvconc(cfb, loess, linear, se_loess, se_linear)
}
\arguments{
\item{cfb}{Logical indicating if dependent variable is a change from baseline.
Plots a reference line at y = cfb_baseline. Default is \code{FALSE}.}

\item{loess}{Logical indicating if a loess smoother fit should be shown. Default is \code{TRUE}}

\item{linear}{Logical indicating if a linear regression fit should be shown. Default is \code{FALSE}.}

\item{se_loess}{Logical indicating if the standard error should be shown for the loess fit. Default is \code{FALSE}}

\item{se_linear}{Logical indicating if the standard error should be shown for the linear fit. Default is \code{FALSE}}
}
\value{
a \code{character} string containing the plot caption
}
\description{
Define a caption for \code{plot_dvconc}
}
\examples{
caption_dvconc(cfb=FALSE, loess = TRUE, linear = FALSE, se_loess = FALSE, se_linear = FALSE)
}
\keyword{internal}
