% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_addn.R
\name{df_addn}
\alias{df_addn}
\title{Add count unique values of an identifier within each level of a variable in a data.frame}
\usage{
df_addn(data, grp_var, id_var = ID, new_var = NULL, sep = NULL)
}
\arguments{
\item{data}{Input dataset.}

\item{grp_var}{Column of values to add counts. Accepts bare names or strings.}

\item{id_var}{Column defining distinct values to count. Accepts bare names or strings. Default is \code{ID}.}

\item{new_var}{Column name for new variable containing values of \code{grp_var} with counts of \code{id_var}.
Default is \code{NULL}, which overwrites \code{grp_var}.}

\item{sep}{Additional string separator to add between variable and count. Default is NULL.}
}
\value{
A data.frame with the same number of rows as \code{data} and a factor variable.
}
\description{
\code{df_addn()} returns a data.frame with a factor variable including count of individuals.
}
\examples{
data <- df_addn(data = data_sad, grp_var = DOSE, id_var = ID)
unique(data$DOSE)

data <- df_addn(data = data_sad, grp_var = DOSE, id_var = ID, sep = "mg")
unique(data$DOSE)

}
