% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_addpred.R
\name{df_addpred}
\alias{df_addpred}
\title{Add population predictions (\code{PRED}) to a data.frame}
\usage{
df_addpred(data, model, output_var = "IPRED", ...)
}
\arguments{
\item{data}{Input dataset.}

\item{model}{\code{mrgsolve} model object.}

\item{output_var}{Name of output from \code{model} to be captured as \code{PRED}
after removing random effects with \code{\link[mrgsolve:zero_re]{mrgsolve::zero_re()}}. Default is \code{"IPRED"}.}

\item{...}{Additional arguments passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}}.}
}
\value{
A data.frame with the same number of rows as \code{data} and on additional
numeric variable \code{PRED}.
}
\description{
\code{df_addpred()} is a wrapper function for
\code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}} and \code{\link[mrgsolve:zero_re]{mrgsolve::zero_re()}} that returns
a data.frame with the addition of a new variable (\code{PRED}).
}
\examples{
model <- model_mread_load(model = "model")
data <- df_addpred(data = data_sad, model = model)

}
