% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_plot_exactbins.R
\name{df_calc_predcorr}
\alias{df_calc_predcorr}
\title{Perform prediction-correction}
\usage{
df_calc_predcorr(
  data,
  bin_var = "NTIME",
  strat_vars = "CMT",
  output_vars = c(PRED = "PRED", IPRED = "IPRED", DV = "DV")
)
}
\arguments{
\item{data}{Input dataset}

\item{bin_var}{Exact binning variable. Default is \code{"NTIME"}.}

\item{strat_vars}{Stratifying variables. Default is \code{"CMT"}.}

\item{output_vars}{Names of model outputs from \code{model}. Must be named character vector.
Defaults are \code{"PRED"}, \code{"IPRED"}, and \code{"DV"}.}
}
\value{
A data.frame containing one row per unique combination of
\code{bin_var} and \code{strat_vars} and new variable \code{PCDV} containing
prediction-corrected observations.
}
\description{
\code{df_calc_predcorr} is a helper function to perform prediction-correction
of observed or simulated depedent variables.
}
\examples{
model <- model_load(model = "model")
data <- df_add_pred(data_sad, model)
simout <- df_calc_predcorr(data, output_vars = c(DV = "ODV"))

}
