% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doseprop.R
\name{df_loglog}
\alias{df_loglog}
\title{Compute estimate table for log-log regression}
\usage{
df_loglog(fit)
}
\arguments{
\item{fit}{\code{lm} model object for the log-log regression}
}
\value{
\code{data.frame}
}
\description{
Compute estimate table for log-log regression
}
\examples{
#example needed
}
