% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mrgsim_addpred.R
\name{df_mrgsim_addpred}
\alias{df_mrgsim_addpred}
\title{Add population predictions (\code{PRED}) to a data.frame}
\usage{
df_mrgsim_addpred(data, model, output_var = IPRED, ...)
}
\arguments{
\item{data}{Input dataset.}

\item{model}{\code{mrgsolve} model object.}

\item{output_var}{Name of output from \code{model} to be captured as \code{PRED}
after removing random effects with \code{\link[mrgsolve:zero_re]{mrgsolve::zero_re()}}. Accepts bare names or strings. Default is \code{IPRED}.}

\item{...}{Additional arguments passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}}.}
}
\value{
A data.frame with the same number of rows as \code{data} and on additional
numeric variable \code{PRED}.
}
\description{
\code{df_mrgsim_addpred()} is a wrapper function for
\code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}} and \code{\link[mrgsolve:zero_re]{mrgsolve::zero_re()}} that returns
a data.frame with the addition of a new variable (\code{PRED}).
}
\examples{
pkmodel <- model_mread_load(model = "pkmodel")
data <- df_mrgsim_addpred(data = dplyr::filter(data_sad, CMT != 3), model = pkmodel)

}
