% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_exactbins.R
\name{df_nobsbin}
\alias{df_nobsbin}
\title{Count the non-missing observations in each exact bin}
\usage{
df_nobsbin(data, bin_var = "NTIME", strat_vars = NULL)
}
\arguments{
\item{data}{Input dataset.}

\item{bin_var}{Binning variable. Default is \code{"NTIME"}.}

\item{strat_vars}{Stratifying variables. Must be a character vector.
Default is \code{"CMT"}.}
}
\value{
A data.frame containing one row per unique combination of
\code{bin_var} and \code{strat_vars} and new variables \code{n_obs}, a count
of non-missing observations, and \code{n_miss}, a count of missing observations.
}
\description{
\code{df_nobsbin()} is a helper function to count the number of missing
and non-missing observations in exact bins.
}
\examples{
df_nobsbin(data_sad)

}
