% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_prep_timevars}
\alias{df_prep_timevars}
\title{Internal helper: Standardize time variable names in a data.frame}
\usage{
df_prep_timevars(data, time_var_str, ntime_var_str)
}
\arguments{
\item{data}{Input data.frame.}

\item{time_var_str}{String name of the actual time column in \code{data}.}

\item{ntime_var_str}{String name of the nominal time column in \code{data}.}
}
\value{
A data.frame with standardized \code{TIME} and \code{NTIME} columns
}
\description{
Renames columns in \code{data} to the standardized \code{TIME} and \code{NTIME} names.
When both time variables map to the same source column, the column is
renamed to \code{NTIME} and \code{TIME} is created as a copy (or vice versa if the
source is already named \code{TIME}).
}
\examples{
data <- dplyr::rename(data_sad, NTFD = NTIME)
c("TIME", "NTIME") \%in\% colnames(data)
data_std <- pmxhelpr:::df_prep_timevars(data, "TIME", "NTFD")
c("TIME", "NTIME") \%in\% colnames(data_std)

}
\keyword{internal}
