% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvtime.R
\name{dvtime_caption}
\alias{dvtime_caption}
\title{Define a caption for \code{plot_dvtime}}
\usage{
dvtime_caption(cent, log_y = FALSE, obs_dv = TRUE, grp_dv = FALSE)
}
\arguments{
\item{cent}{Character string specifying the central tendency measure to plot.

Options are:
\itemize{
\item Mean only: \code{"mean"} (default)
\item Mean +/- Standard Deviation (upper and lower error bar): \code{"mean_sdl"}
\item Mean + Standard Deviation (upper error bar only): \code{"mean_sdl_upper"}
\item Median only: \code{"median"}
\item Median +/- Interquartile Range: \code{median_iqr}
\item None: \code{"none"}
}}

\item{log_y}{Logical indicator for log10 transformation of the y-axis.}

\item{obs_dv}{Logical indicating if observed data points should be shown. Default is \code{TRUE}.}

\item{grp_dv}{Logical indicating if observed data points should be connected within a group (i.e., spaghetti plot).
Default is \code{FALSE}.}
}
\value{
a \code{character} string containing the plot caption
}
\description{
Define a caption for \code{plot_dvtime}
}
\examples{
dvtime_caption(cent = "mean")
dvtime_caption(cent = "mean", log_y = TRUE)
}
