% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_update}
\alias{list_update}
\title{Internal helper: Update elements of a list by name}
\usage{
list_update(update = NULL, src)
}
\arguments{
\item{update}{Named list of values to update
Default is \code{NULL}}

\item{src}{Named list of source values}
}
\value{
A named list
}
\description{
Internal helper: Update elements of a list by name
}
\examples{
src_list <- c(a = 1, b = 2, c = 3, d = 4)
new_list <- pmxhelpr:::list_update(c(a = 5), src_list)

}
\keyword{internal}
