% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doseprop.R
\name{mod_loglog}
\alias{mod_loglog}
\title{Perform a log-log Regression}
\usage{
mod_loglog(data, exp_var = "PPORRES", dose_var = "DOSE")
}
\arguments{
\item{data}{Input dataset for log-log regression.
Default expected format is output from \code{PKNCA::pk.nca()} (i.e., SDTM PP formatting)}

\item{exp_var}{Character string specifying the variable in \code{data} containing the exposure metric (dependent variable)
Default is "PPORRES".}

\item{dose_var}{Character string specifying the variable in \code{data} containing the dose (independent variable)
Default is "DOSE".}
}
\value{
\code{lm} object
}
\description{
Perform a log-log Regression
}
\examples{
#example needed

}
