% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim_vpc.R
\name{mrgsim_vpc}
\alias{mrgsim_vpc}
\title{mrgsim_vpc}
\usage{
mrgsim_vpc(
  data,
  model,
  replicates,
  time_vars = c(TIME = "TIME", NTIME = "NTIME"),
  output_vars = c(PRED = "PRED", IPRED = "IPRED", DV = "DV"),
  num_vars = c("CMT", "EVID", "MDV", "NTIME"),
  char_vars = NULL,
  irep_name = "SIM",
  seed = 123456789,
  ...
)
}
\arguments{
\item{data}{simulation input dataset}

\item{model}{mrgsolve model object}

\item{replicates}{number of iterations in the simulation}

\item{time_vars}{named character vector of time variables (default: c(TIME = "TIME", NTIME = "NTIME"))}

\item{output_vars}{named character vector of numeric output variables to return (default: c(PRED = "PRED", IPRED = "IPRED", DV = "DV"))}

\item{num_vars}{character vector of numeric variable names from the simulation output to return}

\item{char_vars}{character vector of variable names to return}

\item{irep_name}{character string name of the iteration variable (default: "SIM")}

\item{seed}{random seed}

\item{...}{additional arguments passed to mrgsolve::mrgsim_df}
}
\value{
data.frame of simulated output
}
\description{
mrgsim_vpc
}
\examples{
model <- model_load(model = "model")
simout <- mrgsim_vpc(data = data_sad, model = model, replicates = 100,
output_vars = c(DV = "ODV"),
num_vars = c("CMT", "LLOQ", "EVID", "MDV", "WTBL", "FOOD"),
char_vars = c("USUBJID", "PART"),
irep_name = "SIM")
}
