% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvtime_dual.R
\name{plot_dvtime_dual}
\alias{plot_dvtime_dual}
\title{Wrapper function for plot_dvtime to plot two dependent variables paneled together}
\usage{
plot_dvtime_dual(
  data,
  dv_var1 = "DV",
  dv_var2 = "DV",
  dvid_var = "CMT",
  dvid_val1 = 2,
  dvid_val2 = 3,
  time_vars = c(TIME = "TIME", NTIME = "NTIME"),
  timeu = "hours",
  col_var = NULL,
  grp_var = "ID",
  dose_var = "DOSE",
  loq = NULL,
  loq_method = 0,
  cent = "mean",
  obs_dv = TRUE,
  grp_dv = FALSE,
  dosenorm = FALSE,
  cfb = FALSE,
  cfb_base = 0,
  ylab1 = "Concentration",
  ylab2 = "Response",
  log_y1 = FALSE,
  log_y2 = FALSE,
  show_caption = TRUE,
  n_breaks = 8,
  theme = NULL
)
}
\arguments{
\item{data}{Input dataset.}

\item{dv_var1}{Character string specifying the variable containing observations for the top panel (DV1).
Default is "DV".}

\item{dv_var2}{Character string specifying the variable containing observations for the bottom panel (DV2).
Default is "DV".}

\item{dvid_var}{Character string specifying the variable to identify each observation type.}

\item{dvid_val1}{Value of variable specified in \code{dvid_var} to identify observations for top panel (DV1).}

\item{dvid_val2}{Value of variable specified in \code{dvid_var} to identify observations for bottom panel (DV2).}

\item{time_vars}{Names of actual and nominal time variables. Must be named character vector.
Defaults is: c(\code{TIME}=\code{"TIME"}, \code{NTIME}=\code{"NTIME"}).}

\item{timeu}{Character string specifying units for the time variable.
Passed to \code{breaks_time} and assigned to default x-axis label.
Options include:
\itemize{
\item "hours" (default)
\item "days"
\item "weeks"
\item "months"
}}

\item{col_var}{Character string of the name of the variable to map to the color aesthetic.}

\item{grp_var}{Character string of the variable to map to the group aesthetic. Default is \code{"ID"}}

\item{dose_var}{Character string of the variable to use in dosenormalization when \code{dosenorm} = TRUE.
Default is \code{"DOSE"}.}

\item{loq}{Numeric value of the lower limit of quantification (LLOQ) for the assay.
Must be coercible to a numeric if specified. Can be \code{NULL} if variable \code{LLOQ} is present in \code{data}
Specifying this argument implies that \code{DV} is missing in \code{data} where < LLOQ.}

\item{loq_method}{Method for handling data below the lower limit of quantification (BLQ) in the plot.

Options are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{+ `0` : No handling. Plot input dataset `DV` vs `TIME` as is. (default)
+ `1` : Impute all BLQ data at `TIME` <= 0 to 0 and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the linear scale
+ `2` : Impute all BLQ data at `TIME` <= 0 to 1/2 x `loq` and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the log scale where 0 cannot be displayed
}\if{html}{\out{</div>}}}

\item{cent}{Character string specifying the central tendency measure to plot.

Options are:
\itemize{
\item Mean only: \code{"mean"} (default)
\item Mean +/- Standard Deviation (upper and lower error bar): \code{"mean_sdl"}
\item Mean + Standard Deviation (upper error bar only): \code{"mean_sdl_upper"}
\item Median only: \code{"median"}
\item Median +/- Interquartile Range: \code{median_iqr}
\item None: \code{"none"}
}}

\item{obs_dv}{Logical indicating if observed data points should be shown. Default is \code{TRUE}.}

\item{grp_dv}{Logical indicating if observed data points should be connected within a group (i.e., spaghetti plot).
Default is \code{FALSE}.}

\item{dosenorm}{logical indicating if observed data points should be dose normalized. Default is \code{FALSE},
Requires variable specified in \code{dose_var} to be present in \code{data}}

\item{cfb}{Logical indicating if dependent variable is a change from baseline.
Plots a reference line at y = cfb_baseline. Default is \code{FALSE}.}

\item{cfb_base}{Value for y-intercept when cfb = \code{TRUE}. Default is 0.}

\item{ylab1}{Character string specifying the y-axis label for DV1: Default is \code{"Concentration"}.}

\item{ylab2}{Character string specifying the y-axis label for DV2: Default is \code{"Response"}.}

\item{log_y1}{Logical indicator for log10 transformation of the y-axis for DV1. Default is \code{FALSE}.}

\item{log_y2}{Logical indicator for log10 transformation of the y-axis. Default is \code{FALSE}}

\item{show_caption}{Logical indicating if a caption should be show describing the data plotted}

\item{n_breaks}{Number of breaks requested for x-axis. Default is 8.}

\item{theme}{Named list of aesthetic parameters to be supplied to the plot.
Defaults can be viewed by running \code{plot_dvtime_theme()} with no arguments.
Default \code{width_errorbar} is 2.5\% of maximum \code{NTIME}.}
}
\value{
A \code{ggplot2} plot object
}
\description{
Wrapper function for plot_dvtime to plot two dependent variables paneled together
}
\examples{
data <- df_addn(data_sad_pd, id_var = "ID", grp_var = "DOSE", sep = "mg")
plot_dvtime_dual(data, dv_var1 = "ODV", dv_var2 = "ODV")

}
