% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shown.R
\name{plot_gof_shown}
\alias{plot_gof_shown}
\title{GOF layer visibility settings}
\usage{
plot_gof_shown(obs = NULL, dv = NULL, pred = NULL, ipred = NULL)
}
\arguments{
\item{obs}{Show observed data points/lines. Default is \code{TRUE}.}

\item{dv}{Show DV central tendency. Default is \code{TRUE}.}

\item{pred}{Show PRED central tendency. Default is \code{TRUE}.}

\item{ipred}{Show IPRED central tendency. Default is \code{TRUE}.}
}
\value{
A named list of logicals
}
\description{
Constructor and factory for controlling which GOF overlay layers are displayed.
Call with no arguments to view defaults. Pass overrides to customize.
}
\examples{
plot_gof_shown()
plot_gof_shown(pred = FALSE)
}
\seealso{
Other goodness-of-fit:
\code{\link[=plot_gof]{plot_gof()}},
\code{\link[=style_gof]{style_gof()}}
}
\concept{goodness-of-fit}
