% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_legend.R
\name{plot_legend}
\alias{plot_legend}
\title{Plot a legend for a visual predictive check (VPC)}
\usage{
plot_legend(
  ci = c(0.05, 0.95),
  pi = c(0.05, 0.95),
  shown = NULL,
  lloq = NULL,
  update = NULL,
  ...
)
}
\arguments{
\item{ci}{simulated confidence interval plotted. Should match argument passed to \code{\link[vpc:vpc]{vpc::vpc()}}. Default is c(0.05, 0.95).}

\item{pi}{prediction intervals plotted. Should match argument passed to \code{\link[vpc:vpc]{vpc::vpc()}}. Default is c(0.05, 0.95).}

\item{shown}{Named list of logicals specifying which layers to include on the plot. Passed to \code{show} argument of \code{\link[vpc:vpc]{vpc::vpc()}}.

Defaults are:
\itemize{
\item Observed points: \code{obs_dv} = TRUE.
\item Observed quantiles: \code{obs_ci} = TRUE
\item Simulated inter-quantile range:\code{pi} = FALSE
\item Simulated inter-quantile area: \code{pi_as_area} = FALSE
\item Simulated Quantile CI: \code{pi_ci} = TRUE
\item Observed Median: \code{obs_median} = TRUE
\item Simulated Median: \code{sim_median} = FALSE
\item Simulated Median CI: \code{sim_median_ci} = TRUE
}}

\item{lloq}{label for lower limit of quantification in the plot legend.}

\item{update}{list containing the plot elements to be updated. Default is set by \code{\link[vpc:new_vpc_theme]{vpc::new_vpc_theme()}}.}

\item{...}{Other arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
a ggplot2 object
}
\description{
Plot a legend for a visual predictive check (VPC)
}
\examples{
plot_legend()
plot_legend(
pi = c(0.025, 0.975),
ci = c(0.025, 0.925),
 shown = list(obs_dv = FALSE, obs_ci = TRUE,
 pi = FALSE, pi_as_area = FALSE, pi_ci = TRUE,
 obs_median = TRUE,
 sim_median =FALSE, sim_median_ci = TRUE))
}
