% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popgof.R
\name{plot_popgof}
\alias{plot_popgof}
\title{Plot population overlay goodness-of-fit (GOF) plots}
\usage{
plot_popgof(
  data,
  dv_var = DV,
  pred_var = PRED,
  ipred_var = IPRED,
  time_var = TIME,
  ntime_var = NTIME,
  output_colors = .popgof_colors_default,
  grp_var = ID,
  dose_var = DOSE,
  loq = NULL,
  loq_method = 0,
  cent = "mean",
  obs_dv = TRUE,
  grp_dv = FALSE,
  dosenorm = FALSE,
  cfb = FALSE,
  cfb_base = 0,
  log_y = FALSE,
  show_caption = TRUE,
  theme = NULL
)
}
\arguments{
\item{data}{Input dataset.}

\item{dv_var}{Column containing the dependent variable (DV).
Accepts bare names or strings. Default is \code{DV}.}

\item{pred_var}{Column containing population predictions (PRED).
Accepts bare names or strings. Default is \code{PRED}.}

\item{ipred_var}{Column containing individual predictions (IPRED).
Accepts bare names or strings. Default is \code{IPRED}.}

\item{time_var}{Column containing the actual time variable.
Accepts bare names or strings. Default is \code{TIME}.}

\item{ntime_var}{Column containing the nominal time variable.
Accepts bare names or strings. Default is \code{NTIME}.}

\item{output_colors}{Colors for model outputs. Must be named character vector.

Defaults are:
\itemize{
\item \code{PRED= "red"}
\item \code{IPRED}=\code{"green"},
\item \code{DV}=\code{"blue"}.
\item \code{OBS}=\code{"darkgrey"}
}}

\item{grp_var}{Column to map to the group aesthetic. Accepts bare names or strings. Default is \code{ID}.}

\item{dose_var}{Column to use in dosenormalization when \code{dosenorm} = TRUE.
Accepts bare names or strings. Default is \code{DOSE}.}

\item{loq}{Numeric value of the lower limit of quantification (LLOQ) for the assay.
Must be coercible to a numeric if specified. Can be \code{NULL} if variable \code{LLOQ} is present in \code{data}
Specifying this argument implies that \code{DV} is missing in \code{data} where < LLOQ.}

\item{loq_method}{Method for handling data below the lower limit of quantification (BLQ) in the plot.

Options are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{+ `0` : No handling. Plot input dataset `DV` vs `TIME` as is. (default)
+ `1` : Impute all BLQ data at `TIME` <= 0 to 0 and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the linear scale
+ `2` : Impute all BLQ data at `TIME` <= 0 to 1/2 x `loq` and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the log scale where 0 cannot be displayed
}\if{html}{\out{</div>}}}

\item{cent}{Character string specifying the central tendency measure to plot.

Options are:
\itemize{
\item Mean only: \code{"mean"} (default)
\item Mean +/- Standard Deviation (upper and lower error bar): \code{"mean_sdl"}
\item Mean + Standard Deviation (upper error bar only): \code{"mean_sdl_upper"}
\item Median only: \code{"median"}
\item Median +/- Interquartile Range: \code{median_iqr}
\item None: \code{"none"}
}}

\item{obs_dv}{Logical indicating if observed data points should be shown. Default is \code{TRUE}.}

\item{grp_dv}{Logical indicating if observed data points should be connected within a group (i.e., spaghetti plot).
Default is \code{FALSE}.}

\item{dosenorm}{logical indicating if observed data points should be dose normalized. Default is \code{FALSE},
Requires variable specified in \code{dose_var} to be present in \code{data}}

\item{cfb}{Logical indicating if dependent variable is a change from baseline.
Plots a reference line at y = cfb_baseline. Default is \code{FALSE}.}

\item{cfb_base}{Value for y-intercept when cfb = \code{TRUE}. Default is 0.}

\item{log_y}{Logical indicator for log10 transformation of the y-axis. Also controls whether
the caption reports arithmetic or geometric mean when \code{show_caption = TRUE}.}

\item{show_caption}{Logical indicating if a caption should be show describing the data plotted}

\item{theme}{Theme object created by \code{\link[=plot_popgof_theme]{plot_popgof_theme()}}.
Defaults can be viewed by running \code{plot_popgof_theme()} with no arguments.
Default error bar width is 2.5\% of maximum \code{NTIME}.}
}
\value{
A \code{ggplot2} plot object
}
\description{
Plot population overlay goodness-of-fit (GOF) plots
}
\examples{
plot_popgof(data_sad_pkfit, dv_var = ODV, dosenorm = TRUE)

}
