% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{plot_popgof_theme}
\alias{plot_popgof_theme}
\title{Population overlay GOF plot theme}
\usage{
plot_popgof_theme(
  obs = NULL,
  ref = NULL,
  cent = NULL,
  errorbar = NULL,
  dv_line = NULL
)
}
\arguments{
\item{obs}{Observed data aesthetics. See \code{\link[=pmx_obs]{pmx_obs()}}.}

\item{ref}{Reference line aesthetics. See \code{\link[=pmx_ref]{pmx_ref()}}.}

\item{cent}{Central tendency aesthetics. See \code{\link[=pmx_cent]{pmx_cent()}}.}

\item{errorbar}{Error bar aesthetics. See \code{\link[=pmx_errorbar]{pmx_errorbar()}}.}

\item{dv_line}{DV overlay line aesthetics. See \code{\link[=pmx_dv_line]{pmx_dv_line()}}.}
}
\value{
A named list of theme elements
}
\description{
Constructor and factory for \code{plot_popgof} plot aesthetics.
Call with no arguments to view defaults. Pass element overrides to customize.
}
\examples{
plot_popgof_theme()
plot_popgof_theme(obs = pmx_obs(size = 2))
}
