% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_plot_databins.R
\name{plot_vpc_obs_actual}
\alias{plot_vpc_obs_actual}
\title{plot_vpc_obs_actual}
\usage{
plot_vpc_obs_actual(
  sim,
  pcvpc = FALSE,
  strat_vars = NULL,
  irep_name = "SIM",
  min_bin_count = 1,
  ...
)
}
\arguments{
\item{sim}{simulation output data.frame (currently requires variable TIME))}

\item{pcvpc}{logical for prediction correction (default = FALSE)}

\item{strat_vars}{character vector (max length = 2) of stratifying variables}

\item{irep_name}{name of the iteration variable (default = SIM)}

\item{min_bin_count}{minimum number of quantifiable observations per bin for plotting.
Drops small bins from summary statistic calculation but retains in observed data points.}

\item{...}{other arguments passed to vpc::vpc}
}
\value{
ggplot2 object
}
\description{
plot_vpc_obs_actual
}
\examples{
#need examples
}
