% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_exactbins.R
\name{plot_vpc_shown}
\alias{plot_vpc_shown}
\title{Default VPC show layer settings}
\usage{
plot_vpc_shown(update = NULL)
}
\arguments{
\item{update}{list containing the plot elements to be updated.
Run \code{plot_vpc_shown()} with no arguments to view defaults.}
}
\value{
a \code{list}with vpc shown specifiers
}
\description{
Default VPC show layer settings
}
\examples{
plot_vpc_shown()
}
