% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{plot_vpc_theme}
\alias{plot_vpc_theme}
\title{VPC plot theme}
\usage{
plot_vpc_theme(
  obs = NULL,
  obs_median = NULL,
  obs_ci = NULL,
  sim_pi = NULL,
  sim_median = NULL,
  loq = NULL,
  bin_sep = NULL
)
}
\arguments{
\item{obs}{Observed data point aesthetics. See \code{\link[=pmx_vpc_point]{pmx_vpc_point()}}.}

\item{obs_median}{Observed median line aesthetics. See \code{\link[=pmx_vpc_line]{pmx_vpc_line()}}.}

\item{obs_ci}{Observed CI line aesthetics. See \code{\link[=pmx_vpc_line]{pmx_vpc_line()}}.}

\item{sim_pi}{Simulated prediction interval ribbon aesthetics. See \code{\link[=pmx_vpc_ribbon]{pmx_vpc_ribbon()}}.}

\item{sim_median}{Simulated median ribbon aesthetics. See \code{\link[=pmx_vpc_ribbon]{pmx_vpc_ribbon()}}.}

\item{loq}{LOQ reference line aesthetics. See \code{\link[=pmx_vpc_loq]{pmx_vpc_loq()}}.}

\item{bin_sep}{Bin separator aesthetics. See \code{\link[=pmx_vpc_bin_sep]{pmx_vpc_bin_sep()}}.}
}
\value{
A named list of theme elements
}
\description{
Constructor and factory for \code{plot_vpc_cont} plot aesthetics.
Call with no arguments to view defaults. Pass element overrides to customize.
}
\examples{
plot_vpc_theme()
plot_vpc_theme(obs = pmx_vpc_point(color = "#000000"))
}
