% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_constructors.R
\name{pmx_cent}
\alias{pmx_cent}
\title{Central tendency point and line aesthetics}
\usage{
pmx_cent(
  shape = NULL,
  size = NULL,
  alpha = NULL,
  linewidth = NULL,
  linetype = NULL,
  line_alpha = NULL
)
}
\arguments{
\item{shape}{Point shape. Default varies by plot type.}

\item{size}{Point size. Default varies by plot type.}

\item{alpha}{Point alpha. Default varies by plot type.}

\item{linewidth}{Line width. Default varies by plot type.}

\item{linetype}{Line type. Default varies by plot type.}

\item{line_alpha}{Line alpha. Default varies by plot type.}
}
\value{
A \code{pmx_cent} element object
}
\description{
Central tendency point and line aesthetics
}
