% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{pmx_dv_line}
\alias{pmx_dv_line}
\title{DV overlay line aesthetics (popgof only)}
\usage{
pmx_dv_line(linewidth = NULL, linetype = NULL, line_alpha = NULL)
}
\arguments{
\item{linewidth}{Line width. Default is 1.}

\item{linetype}{Line type. Default is 1.}

\item{line_alpha}{Line alpha. Default is 1.}
}
\value{
A \code{pmx_dv_line} element object
}
\description{
DV overlay line aesthetics (popgof only)
}
