% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_constructors.R
\name{pmx_errorbar}
\alias{pmx_errorbar}
\title{Error bar aesthetics}
\usage{
pmx_errorbar(linewidth = NULL, linetype = NULL, alpha = NULL, width = NULL)
}
\arguments{
\item{linewidth}{Error bar line width. Default varies by plot type.}

\item{linetype}{Error bar line type. Default varies by plot type.}

\item{alpha}{Error bar alpha. Default varies by plot type.}

\item{width}{Error bar cap width. Default is 2.5 percent of maximum \code{NTIME}.}
}
\value{
A \code{pmx_errorbar} element object
}
\description{
Error bar aesthetics
}
