% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{pmx_obs}
\alias{pmx_obs}
\title{Observed data point and line aesthetics}
\usage{
pmx_obs(
  shape = NULL,
  size = NULL,
  alpha = NULL,
  linewidth = NULL,
  linetype = NULL,
  line_alpha = NULL
)
}
\arguments{
\item{shape}{Point shape. Default varies by plot type.}

\item{size}{Point size. Default varies by plot type.}

\item{alpha}{Point alpha. Default varies by plot type.}

\item{linewidth}{Spaghetti line width. Default varies by plot type.}

\item{linetype}{Spaghetti line type. Default varies by plot type.}

\item{line_alpha}{Spaghetti line alpha. Default varies by plot type.}
}
\value{
A \code{pmx_obs} element object
}
\description{
Observed data point and line aesthetics
}
