% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_constructors.R
\name{pmx_ref}
\alias{pmx_ref}
\title{Reference line aesthetics}
\usage{
pmx_ref(linewidth = NULL, linetype = NULL, alpha = NULL)
}
\arguments{
\item{linewidth}{Line width. Default varies by plot type.}

\item{linetype}{Line type. Default varies by plot type.}

\item{alpha}{Line alpha. Default varies by plot type.}
}
\value{
A \code{pmx_ref} element object
}
\description{
Reference line aesthetics
}
