% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_element_class.R
\name{pmx_ribbon}
\alias{pmx_ribbon}
\title{Ribbon aesthetics}
\usage{
pmx_ribbon(
  fill = NULL,
  alpha = NULL,
  color = NULL,
  linetype = NULL,
  linewidth = NULL
)
}
\arguments{
\item{fill}{Ribbon fill color. Default varies by ribbon type.}

\item{alpha}{Ribbon alpha. Default varies by ribbon type.}

\item{color}{Ribbon border color. Default varies by ribbon type.}

\item{linetype}{Ribbon border line type. Default varies by ribbon type.}

\item{linewidth}{Ribbon border line width. Default varies by ribbon type.}
}
\value{
A \code{pmx_ribbon} element object
}
\description{
Constructor for ribbon layer aesthetics. Used for simulated prediction
interval and median CI ribbons in VPC plots.
}
\examples{
pmx_ribbon(fill = "skyblue", alpha = 0.3)
}
\seealso{
Other element constructors:
\code{\link[=pmx_color]{pmx_color()}},
\code{\link[=pmx_errorbar]{pmx_errorbar()}},
\code{\link[=pmx_line]{pmx_line()}},
\code{\link[=pmx_point]{pmx_point()}},
\code{\link[=pmx_style]{pmx_style()}},
\code{\link[=pmx_trend]{pmx_trend()}}
}
\concept{element constructors}
