% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_constructors.R
\name{pmx_trend}
\alias{pmx_trend}
\title{Trend line aesthetics (dvconc loess/linear)}
\usage{
pmx_trend(
  linewidth = NULL,
  linetype = NULL,
  color = NULL,
  se_color = NULL,
  se_alpha = NULL
)
}
\arguments{
\item{linewidth}{Line width. Default varies by trend type.}

\item{linetype}{Line type. Default varies by trend type.}

\item{color}{Line color. Default is \code{"black"}.}

\item{se_color}{Standard error ribbon color. Default is \code{"lightgrey"}.}

\item{se_alpha}{Standard error ribbon alpha. Default is 0.4.}
}
\value{
A \code{pmx_trend} element object
}
\description{
Trend line aesthetics (dvconc loess/linear)
}
