% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{pmx_vpc_bin_sep}
\alias{pmx_vpc_bin_sep}
\title{VPC bin separator aesthetics}
\usage{
pmx_vpc_bin_sep(color = NULL)
}
\arguments{
\item{color}{Line color. Default is \code{"#000000"}.}
}
\value{
A \code{pmx_vpc_bin_sep} element object
}
\description{
VPC bin separator aesthetics
}
