% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{pmx_vpc_line}
\alias{pmx_vpc_line}
\title{VPC line aesthetics}
\usage{
pmx_vpc_line(color = NULL, linetype = NULL, size = NULL)
}
\arguments{
\item{color}{Line color. Default varies by line type.}

\item{linetype}{Line type. Default varies by line type.}

\item{size}{Line size. Default varies by line type.}
}
\value{
A \code{pmx_vpc_line} element object
}
\description{
Used for observed median and CI lines.
}
