% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{pmx_vpc_loq}
\alias{pmx_vpc_loq}
\title{VPC LOQ reference line aesthetics}
\usage{
pmx_vpc_loq(color = NULL, linetype = NULL)
}
\arguments{
\item{color}{Line color. Default is \code{"#990000"}.}

\item{linetype}{Line type. Default is \code{"dashed"}.}
}
\value{
A \code{pmx_vpc_loq} element object
}
\description{
VPC LOQ reference line aesthetics
}
