% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{pmx_vpc_point}
\alias{pmx_vpc_point}
\title{VPC observed data point aesthetics}
\usage{
pmx_vpc_point(color = NULL, size = NULL, shape = NULL, alpha = NULL)
}
\arguments{
\item{color}{Point color. Default is \code{"#0000FF"}.}

\item{size}{Point size. Default is 1.}

\item{shape}{Point shape. Default is 1.}

\item{alpha}{Point alpha. Default is 0.7.}
}
\value{
A \code{pmx_vpc_point} element object
}
\description{
VPC observed data point aesthetics
}
