% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_constructors.R
\name{pmx_vpc_ribbon}
\alias{pmx_vpc_ribbon}
\title{VPC ribbon aesthetics}
\usage{
pmx_vpc_ribbon(
  fill = NULL,
  alpha = NULL,
  color = NULL,
  linetype = NULL,
  size = NULL
)
}
\arguments{
\item{fill}{Ribbon fill color. Default varies by ribbon type.}

\item{alpha}{Ribbon alpha. Default varies by ribbon type.}

\item{color}{Ribbon border color. Default varies by ribbon type.}

\item{linetype}{Ribbon border line type. Default varies by ribbon type.}

\item{size}{Ribbon border size. Default varies by ribbon type.}
}
\value{
A \code{pmx_vpc_ribbon} element object
}
\description{
Used for simulated prediction interval and median CI ribbons.
}
