% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_exactbins.R
\name{pmxhelpr_vpc_theme}
\alias{pmxhelpr_vpc_theme}
\title{Customized VPC theme with pmxhelpr default aesthetics}
\usage{
pmxhelpr_vpc_theme(update = NULL)
}
\arguments{
\item{update}{list containing the plot elements to be updated.
Run \code{pmxhelpr_vpc_theme()} with no arguments to view defaults.}
}
\value{
a \code{list}with vpc theme specifiers
}
\description{
Customized VPC theme with pmxhelpr default aesthetics
}
\examples{
pmxhelpr_vpc_theme()
new_theme <- pmxhelpr_vpc_theme(update = vpc::new_vpc_theme()) #restores vpc package defaults
}
