% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{restyle_plot}
\alias{reveal}
\alias{combine_styled_plots}
\alias{legend_spec}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggstylekit}{\code{\link[ggstylekit:combine_styled_plots]{combine_styled_plots()}}, \code{\link[ggstylekit:legend_spec]{legend_spec()}}, \code{\link[ggstylekit:restyle_plot]{restyle_plot()}}, \code{\link[ggstylekit:reveal]{reveal()}}}
}}

