% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_var}
\alias{resolve_var}
\title{Internal helper: Resolve function input conflicts to standard internal string constants}
\usage{
resolve_var(quo, nullable = FALSE)
}
\arguments{
\item{quo}{Function argument as a quosure (wrapped in rlang::ensym)}

\item{nullable}{Logical argument specifying if the argument may be \code{NULL}.
Default is \code{FALSE}.}
}
\value{
A resolved string constant variable
}
\description{
Internal helper: Resolve function input conflicts to standard internal string constants
}
\examples{
f <- function(x = DV) pmxhelpr:::resolve_var(rlang::enquo(x))
f()
f("ODV")
f(CONC)

}
\keyword{internal}
