% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim_vpc.R
\name{sim_vpc}
\alias{sim_vpc}
\title{mrgsim_vpc}
\usage{
sim_vpc(
  data,
  model,
  seed = 123456789,
  replicates = 1000,
  output_vars = c(PRED = "PRED", IPRED = "IPRED", DV = "DV"),
  num_vars,
  char_vars,
  irep_name = "SIM",
  ...
)
}
\arguments{
\item{data}{simulation input dataset}

\item{model}{mrgsolve model object}

\item{seed}{random seed}

\item{replicates}{number of iterations in the simulation}

\item{output_vars}{character vector of numeric output variables to return (default: c("IPRED", "DV))}

\item{num_vars}{character vector of numeric variable names from the simulation output to return}

\item{char_vars}{character vector of variable names to return}

\item{irep_name}{character string name of the iteration variable (default: "SIM")}

\item{...}{additional arguments passed to mrgsolve::mrgsim_df}
}
\value{
data.frame of simulated output
}
\description{
mrgsim_vpc
}
\examples{
x <- 1

}
