% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_dosenorm}
\alias{var_dosenorm}
\title{Internal Helper: Apply dose-normalization to a variable}
\usage{
var_dosenorm(dv_var, dose_var)
}
\arguments{
\item{dv_var}{Vector containing the dependent variable (DV)}

\item{dose_var}{Vector containing dose}
}
\value{
A numeric vector of dose-normalized values of \code{dv_var}
}
\description{
Internal Helper: Apply dose-normalization to a variable
}
\examples{
data <- dplyr::mutate(data_sad, DNDV = var_dosenorm(ODV, DOSE))
}
\seealso{
Other vectorized helpers:
\code{\link[=var_addn]{var_addn()}},
\code{\link[=var_predcorr]{var_predcorr()}}
}
\concept{vectorized helpers}
