% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{var_loqcens}
\alias{var_loqcens}
\title{Determine left-censoring for quantiles at the lower limit of quantification}
\usage{
var_loqcens(x, p, loq)
}
\arguments{
\item{x}{Vector containing the variable to be censored}

\item{p}{Quantile for computation}

\item{loq}{Numeric value of the lower limit of quantification (LLOQ) for the assay}
}
\value{
Replaces values below loq (including NA) with -Inf, then computes
}
\description{
Determine left-censoring for quantiles at the lower limit of quantification
}
\examples{
data <- data_sad |>
  dplyr::group_by(CMT, NTIME, DOSE) |>
  dplyr::summarize(P05 = pmxhelpr:::var_loqcens(ODV, 0.05, loq = LLOQ),
                   P50 = pmxhelpr:::var_loqcens(ODV, 0.5, loq = LLOQ),
                   P95 = pmxhelpr:::var_loqcens(ODV, 0.05, loq = LLOQ), .groups = "drop")
}
\keyword{internal}
