% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_format.R
\name{var_timebreaks}
\alias{var_timebreaks}
\title{Determine axis breaks automatically for time variables}
\usage{
var_timebreaks(x, unit = "hours", n = 8)
}
\arguments{
\item{x}{Numeric vector of times from which to determine breaks}

\item{unit}{Character string for time units.
Options include:
\itemize{
\item "hours" (default), "hrs", "hour", "hr", "h"
\item "days", "dys", "day", "dy", "d"
\item "weeks", "wks", "week", "wk", "w"
\item "months", "mons", "mos", "month", "mo", "m"
}}

\item{n}{Ideal number of axis breaks requested (default = 8). Passed to \code{labeling::extended()}}
}
\value{
A numeric vector of breaks
}
\description{
Determine axis breaks automatically for time variables
}
\examples{
ntimes <- sort(unique(data_sad$NTIME))
breaks <- pmxhelpr:::var_timebreaks(ntimes)
}
\keyword{internal}
