% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_build_plot.R
\name{vpc_build_plot}
\alias{vpc_build_plot}
\title{Build a VPC ggplot from summary statistics}
\usage{
vpc_build_plot(
  vpcstats,
  bin_var = "BIN_MID",
  strat_var_str = NULL,
  shown = NULL,
  vpc_theme = NULL,
  loq = NULL
)
}
\arguments{
\item{vpcstats}{Data.frame of simulated quantile statistics from \code{df_vpcstats()}.}

\item{bin_var}{String. Binning variable name. Default is \code{"BIN_MID"}.}

\item{strat_var_str}{String or \code{NULL}. Stratification variable name.}

\item{shown}{Named list of logicals specifying which layers to include.}

\item{vpc_theme}{Named list of aesthetic parameters (colors, sizes, etc.).}

\item{loq}{Numeric value for LLOQ reference line or \code{NULL}.}
}
\value{
A ggplot2 object.
}
\description{
\code{vpc_build_plot()} constructs a ggplot2 VPC plot from pre-computed quantile
summary statistics.
}
\keyword{internal}
