% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_plot_databins.R
\name{vpc_plot_databins}
\alias{vpc_plot_databins}
\title{vpc_plot_databins}
\usage{
vpc_plot_databins(
  sim,
  pcvpc = FALSE,
  strat_vars = NULL,
  irep_name = "SIM",
  min_bin_count = 1,
  ...
)
}
\arguments{
\item{sim}{simulation output data.frame (currently requires variable TIME))}

\item{pcvpc}{logical for prediction correction (default = FALSE)}

\item{strat_vars}{character vector (max length = 2) of stratifying variables}

\item{irep_name}{name of the iteration variable (default = SIM)}

\item{min_bin_count}{minimum number of quantifiable observations per bin for plotting.
Drops small bins from summary statistic calculation but retains in observed data points.}

\item{...}{other arguments passed to vpc::vpc}
}
\value{
ggplot2 object
}
\description{
vpc_plot_databins
}
\examples{
model <- model_load(model = "model")
simout <- mrgsim_vpc(data = data_sad, model = model, replicates = 100,
output_vars = c(DV = "ODV"),
num_vars = c("CMT", "EVID", "MDV", "NTIME", "LLOQ", "WTBL", "FOOD"),
char_vars = c("USUBJID", "PART"),
irep_name = "SIM")

vpc_plot <- vpc_plot_databins(
sim = simout,
pcvpc = TRUE,
pi = c(0.05, 0.95),
ci = c(0.05, 0.95),
log_y = TRUE)
}
