% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\docType{data}
\name{vpc_show_defaults}
\alias{vpc_show_defaults}
\title{Default VPC show layer settings}
\format{
A named list with elements:
\describe{
\item{obs_dv}{Show observed data points. Default \code{TRUE}.}
\item{obs_ci}{Show observed quantile lines. Default \code{TRUE}.}
\item{pi}{Show simulated prediction interval lines. Default \code{FALSE}.}
\item{pi_as_area}{Show simulated prediction interval as shaded area. Default \code{FALSE}.}
\item{pi_ci}{Show simulated quantile CI ribbons. Default \code{TRUE}.}
\item{obs_median}{Show observed median line. Default \code{TRUE}.}
\item{sim_median}{Show simulated median line. Default \code{FALSE}.}
\item{sim_median_ci}{Show simulated median CI ribbon. Default \code{TRUE}.}
}
}
\usage{
vpc_show_defaults
}
\description{
A named list of logicals specifying which VPC plot layers to display by default.
}
\examples{
vpc_show_defaults
}
\keyword{datasets}
