% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvtime.R
\name{breaks_time}
\alias{breaks_time}
\title{Determine axis breaks automatically for time variables}
\usage{
breaks_time(x, unit = "hours", n = 8)
}
\arguments{
\item{x}{Numeric vector of times from which to determine breaks}

\item{unit}{Character string for time units.
Options include:
\itemize{
\item "hours" (default)
\item "days"
\item "weeks"
\item "months"
}}

\item{n}{Ideal number of axis breaks requested (default = 5). Passed to \code{labeling::extended()}}
}
\value{
A numeric vector of breaks
}
\description{
Determine axis breaks automatically for time variables
}
\examples{
ntimes <- sort(unique(data_sad$NTIME))
breaks <- breaks_time(ntimes)


}
