% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_internals.R
\name{compact}
\alias{compact}
\title{Internal helper: Remove NULL entries from a list}
\usage{
compact(x)
}
\arguments{
\item{x}{A named list potentially containing NULL values.}
}
\value{
A list with all NULL entries removed
}
\description{
Internal helper: Remove NULL entries from a list
}
\examples{
pmxhelpr:::compact(list(a = 1, b = NULL, c = 3))

}
\keyword{internal}
