% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_add_pred.R
\name{df_add_pred}
\alias{df_add_pred}
\title{df_add_pred}
\usage{
df_add_pred(data, model)
}
\arguments{
\item{data}{simulation input dataset}

\item{model}{mrgsolve model object}
}
\value{
data.frame with additional variable PRED
}
\description{
df_add_pred
}
