% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doseprop.R
\name{df_doseprop}
\alias{df_doseprop}
\title{Compute and tabulate estimates for log-log regression}
\usage{
df_doseprop(
  data,
  metrics,
  metric_var = "PPTESTCD",
  exp_var = "PPORRES",
  dose_var = "DOSE",
  method = "normal",
  ci = 0.95,
  sigdigits = 3
)
}
\arguments{
\item{data}{Input dataset for log-log regression.
Default expected format is output from \code{PKNCA::pk.nca()} (i.e., SDTM PP formatting)}

\item{metrics}{character vector of exposure metrics in \code{data} to plot}

\item{metric_var}{character string of variable in \code{data} containing the values provided in \code{metrics}.
Default is "PPTESTCD".}

\item{exp_var}{Character string specifying the variable in \code{data} containing the exposure metric (dependent variable)
Default is "PPORRES".}

\item{dose_var}{Character string specifying the variable in \code{data} containing the dose (independent variable)
Default is "DOSE".}

\item{method}{character string specifying the distribution to be used to derived the confidence interval.
Options are "normal" (default) and "tdist"}

\item{ci}{confidence interval to be calculated.
Options are 0.95 (default) and 0.90}

\item{sigdigits}{number of significant digits for rounding}
}
\value{
\code{data.frame}
}
\description{
Compute and tabulate estimates for log-log regression
}
\examples{
df_doseprop(data_sad_nca, metrics = c("aucinf.obs", "cmax"))
}
