% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doseprop.R
\name{df_loglog}
\alias{df_loglog}
\title{Compute estimate table for log-log regression}
\usage{
df_loglog(fit, method = "normal", ci = 0.95, sigdigits = 3)
}
\arguments{
\item{fit}{\code{lm} model object for the log-log regression}

\item{method}{character string specifying the distribution to be used to derived the confidence interval.
Options are "normal" (default) and "tdist"}

\item{ci}{confidence interval to be calculated.
Options are 0.95 (default) and 0.90}

\item{sigdigits}{number of significant digits for rounding}
}
\value{
\code{data.frame}
}
\description{
Compute estimate table for log-log regression
}
\examples{
#example needed
}
