% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mrgsim_replicate.R
\name{df_mrgsim_replicate}
\alias{df_mrgsim_replicate}
\title{Execute a visual predictive check (VPC) simulation using \code{mrgsolve}}
\usage{
df_mrgsim_replicate(
  data,
  model,
  replicates,
  dv_var = DV,
  time_var = TIME,
  ntime_var = NTIME,
  pred_var = PRED,
  ipred_var = IPRED,
  sim_dv_var = DV,
  irep_name = SIM,
  seed = 123456789,
  ...
)
}
\arguments{
\item{data}{Input dataset. Must contain required variables for \code{mrgsim_df()} other than those handled by
other arguments.}

\item{model}{\code{mrgsolve} model object.}

\item{replicates}{Number of replicates. Either an integer, or something coercible to an integer.}

\item{dv_var}{Column containing the DV variable in \code{data}. Accepts bare names or strings.}

\item{time_var}{Column containing the actual time variable.
Accepts bare names or strings. Default is \code{TIME}.}

\item{ntime_var}{Column containing the nominal time variable.
Accepts bare names or strings. Default is \code{NTIME}.}

\item{pred_var}{Name of population prediction output from \code{model}.
Accepts bare names or strings. Default is \code{PRED}.}

\item{ipred_var}{Name of individual prediction output from \code{model}.
Accepts bare names or strings. Default is \code{IPRED}.}

\item{sim_dv_var}{Name of simulated DV output from \code{model}.
Accepts bare names or strings. Default is \code{DV}.}

\item{irep_name}{Name of replicate variable in \code{data}. Accepts bare names or strings. Default is \code{SIM}.}

\item{seed}{Random seed. Default is \code{123456789}.}

\item{...}{Additional arguments passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}}, including
\code{carry_out} and \code{recover} to control which input columns are propagated to
the output. The always-carried set (\code{EVID}, \code{MDV}, \code{CMT}, \code{TIME}, \code{NTIME},
\code{OBSDV}, and the population prediction column) is added to whatever the
user passes to \code{carry_out}.}
}
\value{
A data.frame with \code{data} x \code{replicates} rows (unless \code{obsonly=TRUE} is passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}})
and the output variables \code{PRED}, \code{IPRED}, \code{SIMDV}, \code{OBSDV}, plus any input
columns listed in \code{carry_out} / \code{recover}.
}
\description{
\code{df_mrgsim_replicate()} is a wrapper function for \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}}
that returns a data.frame containing \code{replicates} iterations of \code{data}
}
\examples{
model <- model_mread_load(model = "pkmodel")
data_sad_pk <- dplyr::filter(data_sad, CMT \%in\% c(1,2))
simout <- df_mrgsim_replicate(data = data_sad_pk, model = model, replicates = 100,
dv_var = ODV,
carry_out = c("LLOQ", "WTBL", "FOOD"),
recover = c("USUBJID", "PART"),
irep_name = SIM)
}
\seealso{
Other mrgsolve wrappers: 
\code{\link{df_mrgsim_addpred}()},
\code{\link{model_mread_load}()}
}
\concept{mrgsolve wrappers}
