% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_prep_blq}
\alias{df_prep_blq}
\title{Internal helper: Apply BLQ censoring rules to data}
\usage{
df_prep_blq(data, loq, loq_method, extra_vars = NULL)
}
\arguments{
\item{data}{data.frame containing data to censor}

\item{loq}{Numeric value of the lower limit of quantification (LLOQ) for the assay.
Must be coercible to a numeric if specified. Can be \code{NULL} if variable \code{LLOQ} is present in \code{data}
Specifying this argument implies that \code{DV} is missing in \code{data} where < LLOQ.}

\item{loq_method}{Method for handling data below the lower limit of quantification (BLQ) in the plot.

Options are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{+ `0` : No handling. Plot input dataset `DV` vs `TIME` as is. (default)
+ `1` : Impute all BLQ data at `TIME` <= 0 to 0 and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the linear scale
+ `2` : Impute all BLQ data at `TIME` <= 0 to 1/2 x `loq` and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the log scale where 0 cannot be displayed
}\if{html}{\out{</div>}}}

\item{extra_vars}{Other variables to include in censoring.}
}
\value{
A data.frame with BLQ censoring applied
}
\description{
Internal helper: Apply BLQ censoring rules to data
}
\examples{
data <- dplyr::rename(dplyr::filter(data_sad, CMT \%in\% c(1,2)), DV = ODV)
data_loq1 <- pmxhelpr:::df_prep_blq(data, loq = 10, loq_method = 1)
data_loq2 <- pmxhelpr:::df_prep_blq(data, loq = 10, loq_method = 2)

}
\keyword{internal}
