% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_load.R
\name{model_load}
\alias{model_load}
\title{model_load}
\usage{
model_load(model, ...)
}
\arguments{
\item{model}{model file name}

\item{...}{additional arguments passed to mrgsolve::mread_cache}
}
\value{
mrgsolve model object
}
\description{
model_load
}
\examples{
model <- model_load("model")
}
