% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mread_load.R
\name{model_mread_load}
\alias{model_mread_load}
\title{Load an mrgsolve model file from the internal model library}
\usage{
model_mread_load(model, ...)
}
\arguments{
\item{model}{Model file name. Must be a string.}

\item{...}{Additional arguments passed to \code{\link[mrgsolve:mread]{mrgsolve::mread_cache()}}.}
}
\value{
An \code{mrgsolve} model object.
}
\description{
\code{model_mread_load()} returns an mrgsolve model object from the internal
model library in \code{pmxhelpr}
}
\examples{
model <- model_mread_load("model")
}
