% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim_vpc.R
\name{mrgsim_vpc}
\alias{mrgsim_vpc}
\title{Execute a visual predictive check (VPC) simulation using \code{mrgsolve}}
\usage{
mrgsim_vpc(
  data,
  model,
  replicates,
  time_vars = c(TIME = "TIME", NTIME = "NTIME"),
  output_vars = c(PRED = "PRED", IPRED = "IPRED", DV = "DV"),
  num_vars = NULL,
  char_vars = NULL,
  irep_name = "SIM",
  seed = 123456789,
  ...
)
}
\arguments{
\item{data}{Input dataset. Must contain required variables for \code{mrgsim_df()} other than those handled by
other arguments.}

\item{model}{\code{mrgsolve} model object.}

\item{replicates}{Number of replicates. Either an integer, or something coercible to an integer.}

\item{time_vars}{Names of actual and nominal time variables. Must be named character vector.
Defaults are \code{"TIME"} and \code{"NTIME"}.}

\item{output_vars}{Names of model outputs from \code{model}. Must be named character vector.
Defaults are \code{"PRED"}, \code{"IPRED"}, and \code{"DV"}.}

\item{num_vars}{Numeric variables in \code{data} or simulation output to recover.
Must be a character vector of variable names from the simulation output to \code{carry_out}
and return in output. Defaults are \code{"CMT"}, \code{"EVID"}, \code{"MDV"}, \code{"NTIME"}.}

\item{char_vars}{Character variables in \code{data} or simulation output to recover.
Must be a character vector of variable names from the simulation output to \code{recover}
and return in output.}

\item{irep_name}{Name of replicate variable in \code{data}. Must be a string. Default is \code{"SIM"}.}

\item{seed}{Random seed. Default is \code{123456789}.}

\item{...}{Additional arguments passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}}.}
}
\value{
A data.frame with \code{data} x \code{replicates} rows (unless \code{obsonly=TRUE})
and the output variables in \code{output_vars}, \code{num_vars}, and \code{char_vars}.
}
\description{
\code{mrgsim_vpc()} is a wrapper function for \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim_df()}}
that returns a data.frame containing \code{replicates} iterations of \code{data}
}
\examples{
model <- model_load(model = "model")
simout <- mrgsim_vpc(data = data_sad, model = model, replicates = 100,
output_vars = c(DV = "ODV"),
num_vars = c("CMT", "LLOQ", "EVID", "MDV", "WTBL", "FOOD"),
char_vars = c("USUBJID", "PART"),
irep_name = "SIM")
}
