% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doseprop.R
\name{plot_doseprop}
\alias{plot_doseprop}
\title{Plot a dose-proportionality assessment via power law (log-log) regression}
\usage{
plot_doseprop(
  data,
  metrics,
  metric_var = "PPTESTCD",
  exp_var = "PPORRES",
  dose_var = "DOSE",
  method = "normal",
  ci = 0.95,
  sigdigits = 3,
  se = TRUE
)
}
\arguments{
\item{data}{Input dataset for log-log regression.
Default expected format is output from \code{PKNCA::pk.nca()} (i.e., SDTM PP formatting)}

\item{metrics}{character vector of exposure metrics in \code{data} to plot}

\item{metric_var}{character string of variable in \code{data} containing the values provided in \code{metrics}.
Default is "PPTESTCD".}

\item{exp_var}{Character string specifying the variable in \code{data} containing the exposure metric (dependent variable)
Default is "PPORRES".}

\item{dose_var}{Character string specifying the variable in \code{data} containing the dose (independent variable)
Default is "DOSE".}

\item{method}{character string specifying the distribution to be used to derived the confidence interval.
Options are "normal" (default) and "tdist"}

\item{ci}{confidence interval to be calculated.
Options are 0.95 (default) and 0.90}

\item{sigdigits}{number of significant digits for rounding}

\item{se}{logical to display confidence interval around regression. Default is \code{TRUE}.}
}
\value{
a \code{ggplot} plot object
}
\description{
Plot a dose-proportionality assessment via power law (log-log) regression
}
\examples{
plot_doseprop(dplyr::filter(data_sad_nca, PART == "Part 1-SAD"), metrics = c("aucinf.obs", "cmax"))
}
