% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvconc.R
\name{plot_dvconc}
\alias{plot_dvconc}
\title{Plot a dependent variable versus concentration}
\usage{
plot_dvconc(
  data,
  dv_var = "DV",
  idv_var = "CONC",
  col_var = NULL,
  col_trend = FALSE,
  loess = TRUE,
  linear = FALSE,
  se_loess = FALSE,
  se_linear = FALSE,
  cfb = FALSE,
  ylab = "Response",
  xlab = "Drug Concentration",
  log_y = FALSE,
  log_x = FALSE,
  show_caption = TRUE,
  theme = NULL,
  ...
)
}
\arguments{
\item{data}{Input dataset.}

\item{dv_var}{Character name of the DV variable in \code{data}.}

\item{idv_var}{Independent variable. Default is \code{"CONC"}.}

\item{col_var}{Character string of the name of the variable to map to the color aesthetic.}

\item{col_trend}{Logical indicating if the variable specified in \code{col_var} should be used to stratify trend lines}

\item{loess}{Logical indicating if a loess smoother fit should be shown. Default is \code{TRUE}}

\item{linear}{Logical indicating if a linear regression fit should be shown. Default is \code{FALSE}.}

\item{se_loess}{Logical indicating if the standard error should be shown for the loess fit. Default is \code{FALSE}}

\item{se_linear}{Logical indicating if the standard error should be shown for the linear fit. Default is \code{FALSE}}

\item{cfb}{Logical indicating if dependent variable is a change from baseline.
Plots a reference line at y = cfb_baseline. Default is \code{FALSE}.}

\item{ylab}{Character string specifying the y-axis label: Default is \code{"Response"}.}

\item{xlab}{Character string specifying the x-axis label: Default is \code{"Drug Concentration"}}

\item{log_y}{Logical indicator for log10 transformation of the y-axis.}

\item{log_x}{Logical indicator for log10 transformation of the x-axis.}

\item{show_caption}{Logical indicating if a caption should be show describing the data plotted}

\item{theme}{Named list of aesthetic parameters to be supplied to the plot.
Defaults can be viewed by running \code{plot_dvconc_theme()} with no arguments.}

\item{...}{Additional arguments passed to \code{geom_smooth()}}
}
\value{
A \code{ggplot2} plot object
}
\description{
Plot a dependent variable versus concentration
}
\examples{
data <- dplyr::mutate(data_sad_pd, Dose = factor(DOSE))
plot_dvconc(data, dv_var = "ODV", idv_var = "ICONC", col_var = "Dose", col_trend = FALSE)

}
