% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{plot_dvconc_theme}
\alias{plot_dvconc_theme}
\title{Response versus concentration plot theme}
\usage{
plot_dvconc_theme(
  obs_point = NULL,
  ref_line = NULL,
  loess = NULL,
  linear = NULL
)
}
\arguments{
\item{obs_point}{Observed data point aesthetics. See \code{\link[=pmx_point]{pmx_point()}}.}

\item{ref_line}{Reference line aesthetics. See \code{\link[=pmx_line]{pmx_line()}}.}

\item{loess}{LOESS trend line aesthetics. See \code{\link[=pmx_trend]{pmx_trend()}}.}

\item{linear}{Linear trend line aesthetics. See \code{\link[=pmx_trend]{pmx_trend()}}.}
}
\value{
A named list of theme elements
}
\description{
Constructor and factory for \code{plot_dvconc} plot aesthetics.
Call with no arguments to view defaults. Pass element overrides to customize.
}
\examples{
plot_dvconc_theme()
plot_dvconc_theme(loess = pmx_trend(color = "red"))
}
