% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvconc.R
\name{plot_dvconc_theme}
\alias{plot_dvconc_theme}
\title{Customized Response versus drug concentration theme with pmxhelpr default aesthetics}
\usage{
plot_dvconc_theme(update = NULL)
}
\arguments{
\item{update}{list containing the plot elements to be updated.
Run \code{plot_dvconc_theme()} with no arguments to view defaults.}
}
\value{
a named list \code{list}
}
\description{
Customized Response versus drug concentration theme with pmxhelpr default aesthetics
}
\examples{
plot_dvconc_theme()
new_theme <- plot_dvconc_theme(update = list(linewidth_ref = 1))
}
