% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvtime.R
\name{plot_dvtime}
\alias{plot_dvtime}
\title{Plot a dependent variable versus time}
\usage{
plot_dvtime(
  data,
  dv_var = c(DV = "DV"),
  time_vars = c(TIME = "TIME", NTIME = "NTIME"),
  timeu = "hours",
  col_var = NULL,
  grp_var = "ID",
  loq = NULL,
  loq_method = 0,
  cent = "mean",
  obs_dv = TRUE,
  grp_dv = FALSE,
  dosenorm = FALSE,
  cfb = FALSE,
  ylab = "Concentration",
  log_y = FALSE,
  show_caption = TRUE,
  n_breaks = 8
)
}
\arguments{
\item{data}{Input dataset. Must contain the variables: \code{"ID"}, \code{"DV"} \code{"MDV"}.}

\item{dv_var}{Character string of the dependent variable. Default is \code{"DV"}.}

\item{time_vars}{Names of actual and nominal time variables. Must be named character vector.
Defaults are \code{"TIME"} and \code{"NTIME"}.}

\item{timeu}{Character string specifying units for the time variable.
Passed to \code{breaks_time} and assigned to default x-axis label.
Options include:
\itemize{
\item "hours" (default)
\item "days"
\item "weeks"
\item "months"
}}

\item{col_var}{Character string of the name of the variable to map to the color aesthetic.}

\item{grp_var}{Character string of the variable to map to the group aesthetic. Default is \code{"ID"}}

\item{loq}{Numeric value of the lower limit of quantification (LLOQ) for the assay.
Must be coercible to a numeric if specified. Can be \code{NULL} if variable \code{LLOQ} is present in \code{data}
Specifying this argument implies that \code{DV}is missing in \code{data} where < LLOQ.}

\item{loq_method}{Method for handling data below the lower limit of quantification (BLQ) in the plot.
Options are:
+ \code{0} : No handling. Plot input dataset \code{DV} vs \code{TIME} as is. (default)
+ \code{1} : Impute all BLQ data at \code{TIME} <= 0 to 0 and all BLQ data at \code{TIME} > 0 to 1/2 x \code{loq}.
Useful for plotting concentration-time data with some data BLQ on the linear scale
+ \code{2} : Impute all BLQ data at \code{TIME} <= 0 to 1/2 x \code{loq} and all BLQ data at \code{TIME} > 0 to 1/2 x \code{loq}.
Useful for plotting concentration-time data with some data BLQ on the log scale where 0 cannot be displayed}

\item{cent}{Character string specifying the central tendency measure to plot.
Options are:
\itemize{
\item Mean only: \code{"mean"} (default)
\item Mean +/- Standard Deviation: \code{"mean_sdl"}
\item Median only: \code{"median"}
\item Median +/- Interquartile Range: \code{median_iqr}
\item None: \code{"none"}
}}

\item{obs_dv}{Logical indicating if observed data points should be shown. Default is \code{TRUE}.}

\item{grp_dv}{Logical indicating if observed data points should be connected within a group (i.e., spaghetti plot).
Default is \code{FALSE}.}

\item{dosenorm}{logical indicating if observed data points should be dose normalized. Default is \code{FALSE},
Requires variable \code{DOSE} to be present in \code{data}}

\item{cfb}{Logical indicating if dependent variable is a change from baseline.
Plots a reference line at y = 0. Default is \code{FALSE}.}

\item{ylab}{Character string specifing the y-axis label: Default is \code{"Concentration"}.}

\item{log_y}{Logical indicator for log10 transformation of the y-axis.}

\item{show_caption}{Logical indicating if a caption should be show describing the data plotted}

\item{n_breaks}{Number of breaks requested for x-axis. Default is 5.}
}
\value{
A ggplot2 plot object
}
\description{
Plot a dependent variable versus time
}
\examples{
data <- dplyr::mutate(data_sad, Dose = factor(DOSE))
plot_dvtime(data, dv_var = c(DV = "ODV"), cent = "median", col_var = "Dose")

}
