% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvtime.R
\name{plot_dvtime}
\alias{plot_dvtime}
\title{Plot a dependent variable versus time}
\usage{
plot_dvtime(
  data,
  dv_var = c(DV = "DV"),
  time_vars = c(TIME = "TIME", NTIME = "NTIME"),
  col_var = NULL,
  loq = NULL,
  loq_method = 0,
  cent = "mean",
  obs_dv = TRUE,
  ind_dv = FALSE,
  cfb = FALSE,
  ylab = "Concentration",
  xlab = "Time",
  log_y = FALSE,
  show_caption = TRUE
)
}
\arguments{
\item{data}{Input dataset. Must contain the variables: \code{"ID"}, \code{"DV"} \code{"MDV"}.}

\item{dv_var}{Character string of the dependent variable. Default is \code{"DV"}.}

\item{time_vars}{Names of actual and nominal time variables. Must be named character vector.
Defaults are \code{"TIME"} and \code{"NTIME"}.}

\item{col_var}{Character string of the name of the variable to map to the color aesthetic.}

\item{loq}{Numeric value of the lower limit of quantification (LLOQ) for the assay. Passed to \code{lloq} argument
of \code{\link[vpc:vpc]{vpc::vpc()}}. Specifying this argument implies that \code{OBSDV}is missing in \code{sim} where < LLOQ.}

\item{loq_method}{Method for handling data below the lower limit of quantification (BLQ) in the plot.
Options are:
+ \code{0} : No handling. Plot input dataset \code{DV} vs \code{TIME} as is. (default)
+ \code{1} : Impute all BLQ data at \code{TIME} <= 0 to 0 and all BLQ data at \code{TIME} > 0 to 1/2 x \code{loq}.
Useful for plotting concentration-time data with some data BLQ on the linear scale
+ \code{2} : Impute all BLQ data at \code{TIME} <= 0 to 1/2 x \code{loq} and all BLQ data at \code{TIME} > 0 to 1/2 x \code{loq}.
Useful for plotting concentration-time data with some data BLQ on the log scale where 0 cannot be displayed}

\item{cent}{Character string specifying the central tendency measure to plot.
Options are:
\itemize{
\item Mean only: \code{"mean"} (default)
\item Mean +/- Standard Deviation: \code{"mean_sdl"}
\item Median only: \code{"median"}
\item None: \code{"none"}
}}

\item{obs_dv}{Logical indicating if observed data points should be shown. Default is \code{TRUE}.}

\item{ind_dv}{Logical indiciating if observed data points shoudld be connected within an individual (i.e., spaghetti plot).
Defaut is \code{FALSE}.}

\item{cfb}{Logical indicating if dependent variable is a change from baseline.
Plots a reference line at y = 0. Default is \code{FALSE}.}

\item{ylab}{Character string specifing the y-axis label: Default is \code{"Concentration"}.}

\item{xlab}{Character string specfing the x-axis label: Default is \code{"Time"}.}

\item{log_y}{Logical indicator for log10 transformation of the y-axis.}

\item{show_caption}{Logical indicating if a caption should be show describing the data plotted}
}
\value{
A ggplot2 plot object
}
\description{
Plot a dependent variable versus time
}
\examples{
data <- dplyr::mutate(data_sad, Dose = factor(DOSE))
plot <- plot_dvtime(data, dv_var = c(DV = "ODV"), cent = "median", col_var = "Dose")
plot


}
