% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{plot_dvtime_theme}
\alias{plot_dvtime_theme}
\title{Concentration-time plot theme}
\usage{
plot_dvtime_theme(obs = NULL, ref = NULL, cent = NULL, errorbar = NULL)
}
\arguments{
\item{obs}{Observed data aesthetics. See \code{\link[=pmx_obs]{pmx_obs()}}.}

\item{ref}{Reference line aesthetics. See \code{\link[=pmx_ref]{pmx_ref()}}.}

\item{cent}{Central tendency aesthetics. See \code{\link[=pmx_cent]{pmx_cent()}}.}

\item{errorbar}{Error bar aesthetics. See \code{\link[=pmx_errorbar]{pmx_errorbar()}}.}
}
\value{
A named list of theme elements
}
\description{
Constructor and factory for \code{plot_dvtime} plot aesthetics.
Call with no arguments to view defaults. Pass element overrides to customize.
}
\examples{
plot_dvtime_theme()
plot_dvtime_theme(obs = pmx_obs(size = 2), ref = pmx_ref(linetype = 3))
}
