% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvtime.R
\name{plot_dvtime_theme}
\alias{plot_dvtime_theme}
\title{Customized Concentration-time theme with pmxhelpr default aesthetics}
\usage{
plot_dvtime_theme(update = NULL)
}
\arguments{
\item{update}{list containing the plot elements to be updated.
Run \code{plot_dvtime_theme()} with no arguments to view defaults.}
}
\value{
a named list \code{list}
}
\description{
Customized Concentration-time theme with pmxhelpr default aesthetics
}
\examples{
plot_dvtime_theme()
new_theme <- plot_dvtime_theme(update = list(linewidth_ref = 1))
}
