% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popgof.R
\name{plot_popgof}
\alias{plot_popgof}
\title{Plot population overlay goodness-of-fit (GOF) plots}
\usage{
plot_popgof(
  data,
  time_vars = c(TIME = "TIME", NTIME = "NTIME"),
  output_vars = c(PRED = "PRED", IPRED = "IPRED", DV = "DV"),
  output_colors = c(PRED = "red", IPRED = "green", DV = "blue", OBS = "darkgrey"),
  timeu = "hours",
  grp_var = "ID",
  dose_var = "DOSE",
  loq = NULL,
  loq_method = 0,
  cent = "mean",
  obs_dv = TRUE,
  grp_dv = FALSE,
  dosenorm = FALSE,
  cfb = FALSE,
  ylab = "Concentration",
  log_y = FALSE,
  show_caption = TRUE,
  n_breaks = 8
)
}
\arguments{
\item{data}{Input dataset. Must contain required variables for \code{mrgsim_df()} other than those handled by
other arguments.}

\item{time_vars}{Names of actual and nominal time variables. Must be named character vector.
Defaults is: c(\code{TIME}=\code{"TIME"}, \code{NTIME}=\code{"NTIME"}).}

\item{output_vars}{Names of model outputs from \code{model}. Must be named character vector.
Defaults is: c(\code{PRED}= \code{"PRED"}, \code{IPRED} = \code{"IPRED"}, \code{DV}= \code{"DV"}).}

\item{output_colors}{Colors for model outputs. Must be named character vector.

Defaults are:
\itemize{
\item \code{PRED= "red"}
\item \code{IPRED}=\code{"green"},
\item \code{DV}=\code{"blue"}.
\item \code{OBS}=\code{"darkgrey"}
}}

\item{timeu}{Character string specifying units for the time variable.
Passed to \code{breaks_time} and assigned to default x-axis label.
Options include:
\itemize{
\item "hours" (default)
\item "days"
\item "weeks"
\item "months"
}}

\item{grp_var}{Character string of the variable to map to the group aesthetic. Default is \code{"ID"}}

\item{dose_var}{Character string of the variable to use in dosenormalization when \code{dosenorm} = TRUE.
Default is \code{"DOSE"}.}

\item{loq}{Numeric value of the lower limit of quantification (LLOQ) for the assay.
Must be coercible to a numeric if specified. Can be \code{NULL} if variable \code{LLOQ} is present in \code{data}
Specifying this argument implies that \code{DV} is missing in \code{data} where < LLOQ.}

\item{loq_method}{Method for handling data below the lower limit of quantification (BLQ) in the plot.

Options are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{+ `0` : No handling. Plot input dataset `DV` vs `TIME` as is. (default)
+ `1` : Impute all BLQ data at `TIME` <= 0 to 0 and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the linear scale
+ `2` : Impute all BLQ data at `TIME` <= 0 to 1/2 x `loq` and all BLQ data at `TIME` > 0 to 1/2 x `loq`.
   Useful for plotting concentration-time data with some data BLQ on the log scale where 0 cannot be displayed
}\if{html}{\out{</div>}}}

\item{cent}{Character string specifying the central tendency measure to plot.

Options are:
\itemize{
\item Mean only: \code{"mean"} (default)
\item Mean +/- Standard Deviation: \code{"mean_sdl"}
\item Median only: \code{"median"}
\item Median +/- Interquartile Range: \code{median_iqr}
\item None: \code{"none"}
}}

\item{obs_dv}{Logical indicating if observed data points should be shown. Default is \code{TRUE}.}

\item{grp_dv}{Logical indicating if observed data points should be connected within a group (i.e., spaghetti plot).
Default is \code{FALSE}.}

\item{dosenorm}{logical indicating if observed data points should be dose normalized. Default is \code{FALSE},
Requires variable specified in \code{dose_var} to be present in \code{data}}

\item{cfb}{Logical indicating if dependent variable is a change from baseline.
Plots a reference line at y = 0. Default is \code{FALSE}.}

\item{ylab}{Character string specifing the y-axis label: Default is \code{"Concentration"}.}

\item{log_y}{Logical indicator for log10 transformation of the y-axis.}

\item{show_caption}{Logical indicating if a caption should be show describing the data plotted}

\item{n_breaks}{Number of breaks requested for x-axis. Default is 5.}
}
\value{
A \code{ggplot2} plot object
}
\description{
Plot population overlay goodness-of-fit (GOF) plots
}
\examples{
plot_popgof(data_sad_pkfit, output_vars = c(DV = "ODV"), dosenorm = TRUE, ylab = "Dose-norm Conc.")

}
