% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popgof.R
\name{plot_popgof_theme}
\alias{plot_popgof_theme}
\title{Customized population overlay goodness-of-fit (GOF) theme with pmxhelpr default aesthetics}
\usage{
plot_popgof_theme(update = NULL)
}
\arguments{
\item{update}{list containing the plot elements to be updated.
Run \code{plot_popgof_theme()} with no arguments to view defaults.}
}
\value{
a named list \code{list}
}
\description{
Customized population overlay goodness-of-fit (GOF) theme with pmxhelpr default aesthetics
}
\examples{
plot_popgof_theme()
new_theme <- plot_popgof_theme(update = list(linewidth_ref = 1))
}
