% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_exactbins.R
\name{plot_vpc_exactbins}
\alias{plot_vpc_exactbins}
\title{Plot a visual predictive check (VPC) with exact time bins}
\usage{
plot_vpc_exactbins(
  sim,
  pcvpc = FALSE,
  time_vars = c(TIME = "TIME", NTIME = "NTIME"),
  output_vars = c(PRED = "PRED", IPRED = "IPRED", SIMDV = "SIMDV", OBSDV = "OBSDV"),
  strat_vars = NULL,
  irep_name = "SIM",
  min_bin_count = 1,
  lower_bound = 0,
  ...
)
}
\arguments{
\item{sim}{Input dataset. Must contain the following variables: \code{"ID"}, \code{TIME}}

\item{pcvpc}{logical for prediction correction. Default is \code{FALSE}.}

\item{time_vars}{Names of actual and nominal time variables. Must be named character vector.
Defaults are \code{"TIME"} and \code{"NTIME"}.}

\item{output_vars}{Names of model outputs from \code{model}. Must be named character vector.
Defaults are \code{"PRED"}, \code{"IPRED"}, and \code{"DV"}.}

\item{strat_vars}{Character vector of stratification variables passed to \code{stratify} argument
of \code{\link[vpc:vpc]{vpc::vpc()}}. Currently, maximum length = 2.}

\item{irep_name}{Name of replicate variable in \code{data}. Must be a string. Default is \code{"SIM"}.}

\item{min_bin_count}{Minimum number of quantifiable observations in exact bin for inclusion
in binned plot layers. This argument drops small bins from summary statistic calculation
but retains these observations in the observed data points.}

\item{lower_bound}{Lower bound of the dependent variable for prediction correction. Default is \code{0}.}

\item{...}{Other arguments passed to \code{\link[vpc:vpc]{vpc::vpc()}}.}
}
\value{
A list containing calculated VPC information (when \code{vpcdb=TRUE}), or a ggplot2 object (default)
}
\description{
\code{plot_vpc_exactbins()} is a wrapper function for \code{\link[vpc:vpc]{vpc::vpc()}}
that returns a \code{ggplot2} object.
}
\examples{
model <- model_mread_load(model = "model")
simout <- df_mrgsim_replicate(data = data_sad, model = model, replicates = 100,
output_vars = c(DV = "ODV"),
num_vars = c("CMT", "EVID", "MDV", "NTIME", "LLOQ", "WTBL", "FOOD"),
char_vars = c("USUBJID", "PART"),
irep_name = "SIM")

vpc_plot <- plot_vpc_exactbins(
sim = simout,
pcvpc = TRUE,
pi = c(0.05, 0.95),
ci = c(0.05, 0.95),
log_y = TRUE)
}
