% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_legend.R
\name{plot_vpc_legend}
\alias{plot_vpc_legend}
\title{Plot a legend for a visual predictive check (VPC)}
\usage{
plot_vpc_legend(
  ci = c(0.05, 0.95),
  pi = c(0.05, 0.95),
  shown = NULL,
  lloq = NULL,
  update = NULL,
  ...
)
}
\arguments{
\item{ci}{simulated confidence interval plotted. Should match argument passed to \code{\link[=plot_vpc_cont]{plot_vpc_cont()}}. Default is c(0.05, 0.95).}

\item{pi}{prediction intervals plotted. Should match argument passed to \code{\link[=plot_vpc_cont]{plot_vpc_cont()}}. Default is c(0.05, 0.95).}

\item{shown}{Layer visibility settings created by \code{\link[=plot_vpc_shown]{plot_vpc_shown()}}.
Defaults can be viewed by running \code{plot_vpc_shown()} with no arguments.}

\item{lloq}{label for lower limit of quantification in the plot legend.}

\item{update}{list containing the plot elements to be updated. Default is set by \code{\link[=plot_vpc_theme]{plot_vpc_theme()}}.}

\item{...}{Other arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
a ggplot2 object
}
\description{
Plot a legend for a visual predictive check (VPC)
}
\examples{
plot_vpc_legend()
plot_vpc_legend(
pi = c(0.025, 0.975),
ci = c(0.025, 0.925),
 shown = plot_vpc_shown(obs_dv = FALSE, obs_ci = TRUE,
 pi = FALSE, pi_as_area = FALSE, pi_ci = TRUE,
 obs_median = TRUE,
 sim_median = FALSE, sim_median_ci = TRUE))
}
