% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_legend.R
\name{plot_vpc_legend}
\alias{plot_vpc_legend}
\title{Plot a legend for a visual predictive check (VPC)}
\usage{
plot_vpc_legend(
  ci = 0.9,
  pi = c(0.05, 0.95),
  shown = NULL,
  lloq = NULL,
  update = NULL,
  ...
)
}
\arguments{
\item{ci}{Numeric confidence level for simulation intervals (e.g., \code{0.90} for 90\% CI).
Should match argument passed to \code{\link[=plot_vpc_cont]{plot_vpc_cont()}}. Default is \code{0.90}.}

\item{pi}{prediction intervals plotted. Should match argument passed to \code{\link[=plot_vpc_cont]{plot_vpc_cont()}}. Default is c(0.05, 0.95).}

\item{shown}{Layer visibility settings created by \code{\link[=plot_vpc_shown]{plot_vpc_shown()}}.
Defaults can be viewed by running \code{plot_vpc_shown()} with no arguments.}

\item{lloq}{label for lower limit of quantification in the plot legend.}

\item{update}{list containing the plot elements to be updated. Default is set by \code{\link[=plot_vpc_theme]{plot_vpc_theme()}}.}

\item{...}{Other arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
a ggplot2 object
}
\description{
Plot a legend for a visual predictive check (VPC)
}
\examples{
plot_vpc_legend()
plot_vpc_legend(
pi = c(0.025, 0.975),
ci = 0.95,
 shown = plot_vpc_shown(obs_point = FALSE, obs_pi_line = TRUE,
 sim_pi_line = FALSE, sim_pi_area = FALSE, sim_pi_ci = TRUE,
 obs_median_line = TRUE,
 sim_median_line = FALSE, sim_median_ci = TRUE))
}
\seealso{
Other vpc: 
\code{\link{df_vpcstats}()},
\code{\link{is_vpc_stats}()},
\code{\link{plot_build_vpc}()},
\code{\link{plot_vpc_cont}()},
\code{\link{plot_vpc_shown}()},
\code{\link{plot_vpc_theme}()}
}
\concept{vpc}
