% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_legend.R
\name{plot_vpc_legend}
\alias{plot_vpc_legend}
\title{Plot a legend for a visual predictive check (VPC)}
\usage{
plot_vpc_legend(ci = c(0.05, 0.95), pi = c(0.05, 0.95))
}
\arguments{
\item{ci}{simulated confidence interval plotted. Should match argument passed to \code{\link[vpc:vpc]{vpc::vpc()}}. Default is c(0.05, 0.95).}

\item{pi}{prediction intervals for plotted. Should match argument passed to \code{\link[vpc:vpc]{vpc::vpc()}}. Default is c(0.05, 0.95),}
}
\value{
a ggplot2 object
}
\description{
Plot a legend for a visual predictive check (VPC)
}
\examples{
plot_vpc_legend()
plot_vpc_legend(pi = c(0.05, 0.95), ci = c(0.025, 0.975))
}
