% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc_exactbins.R
\name{plot_vpc_theme}
\alias{plot_vpc_theme}
\title{Customized VPC theme with pmxhelpr default aesthetics}
\usage{
plot_vpc_theme(update = NULL)
}
\arguments{
\item{update}{list containing the plot elements to be updated.
Run \code{plot_vpc_theme()} with no arguments to view defaults.}
}
\value{
a \code{list}with vpc theme specifiers
}
\description{
Customized VPC theme with pmxhelpr default aesthetics
}
\examples{
plot_vpc_theme()
}
