% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{plot_vpc_theme}
\alias{plot_vpc_theme}
\title{VPC plot theme}
\usage{
plot_vpc_theme(
  obs = NULL,
  obs_median = NULL,
  obs_ci = NULL,
  sim_pi = NULL,
  sim_median = NULL,
  loq_line = NULL
)
}
\arguments{
\item{obs}{Observed data point aesthetics. See \code{\link[=pmx_point]{pmx_point()}}.}

\item{obs_median}{Observed median line aesthetics. See \code{\link[=pmx_line]{pmx_line()}}.}

\item{obs_ci}{Observed CI line aesthetics. See \code{\link[=pmx_line]{pmx_line()}}.}

\item{sim_pi}{Simulated prediction interval ribbon aesthetics. See \code{\link[=pmx_ribbon]{pmx_ribbon()}}.}

\item{sim_median}{Simulated median ribbon aesthetics. See \code{\link[=pmx_ribbon]{pmx_ribbon()}}.}

\item{loq_line}{LOQ reference line aesthetics. See \code{\link[=pmx_line]{pmx_line()}}.}
}
\value{
A named list of theme elements
}
\description{
Constructor and factory for \code{plot_vpc_cont} plot aesthetics.
Call with no arguments to view defaults. Pass element overrides to customize.
}
\examples{
plot_vpc_theme()
plot_vpc_theme(obs = pmx_point(color = "#000000"))
}
