% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{pmx_color}
\alias{pmx_color}
\title{GOF overlay color aesthetics}
\usage{
pmx_color(dv = NULL, pred = NULL, ipred = NULL)
}
\arguments{
\item{dv}{Color for DV central tendency. Default \code{"blue"}.}

\item{pred}{Color for PRED central tendency. Default \code{"red"}.}

\item{ipred}{Color for IPRED central tendency. Default \code{"green"}.}
}
\value{
A \code{pmx_color} element object
}
\description{
Creates a color element for \code{\link[=plot_gof_theme]{plot_gof_theme()}} controlling the manual
color scale for DV, PRED, and IPRED overlay lines.
}
