% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_element_class.R
\name{pmx_color}
\alias{pmx_color}
\title{GOF overlay color aesthetics}
\usage{
pmx_color(dv = NULL, pred = NULL, ipred = NULL)
}
\arguments{
\item{dv}{Color for DV central tendency. Default \code{"blue"}.}

\item{pred}{Color for PRED central tendency. Default \code{"red"}.}

\item{ipred}{Color for IPRED central tendency. Default \code{"green"}.}
}
\value{
A \code{pmx_color} element object
}
\description{
Creates a color element for \code{\link[=plot_gof_theme]{plot_gof_theme()}} controlling the manual
color scale for DV, PRED, and IPRED overlay lines.
}
\examples{
pmx_color(dv = "black", pred = "purple", ipred = "darkgreen")
}
\seealso{
Other element constructors:
\code{\link[=pmx_errorbar]{pmx_errorbar()}},
\code{\link[=pmx_line]{pmx_line()}},
\code{\link[=pmx_point]{pmx_point()}},
\code{\link[=pmx_ribbon]{pmx_ribbon()}},
\code{\link[=pmx_style]{pmx_style()}},
\code{\link[=pmx_trend]{pmx_trend()}}
}
\concept{element constructors}
