% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_element_class.R
\name{pmx_errorbar}
\alias{pmx_errorbar}
\title{Error bar aesthetics}
\usage{
pmx_errorbar(
  linewidth = NULL,
  linetype = NULL,
  alpha = NULL,
  color = NULL,
  width = NULL
)
}
\arguments{
\item{linewidth}{Error bar line width. Default varies by plot type.}

\item{linetype}{Error bar line type. Default varies by plot type.}

\item{alpha}{Error bar alpha. Default varies by plot type.}

\item{color}{Error bar color. Default is \code{NULL} (inherits from ggplot2 default).}

\item{width}{Error bar cap width. Default is 2.5 percent of maximum \code{NTIME}.}
}
\value{
A \code{pmx_errorbar} element object
}
\description{
Error bar aesthetics
}
\examples{
pmx_errorbar(linewidth = 0.5, width = 0.5)
}
\seealso{
Other element constructors:
\code{\link[=pmx_color]{pmx_color()}},
\code{\link[=pmx_line]{pmx_line()}},
\code{\link[=pmx_point]{pmx_point()}},
\code{\link[=pmx_ribbon]{pmx_ribbon()}},
\code{\link[=pmx_style]{pmx_style()}},
\code{\link[=pmx_trend]{pmx_trend()}}
}
\concept{element constructors}
