% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_element_class.R
\name{pmx_line}
\alias{pmx_line}
\title{Line aesthetics}
\usage{
pmx_line(linewidth = NULL, linetype = NULL, alpha = NULL, color = NULL)
}
\arguments{
\item{linewidth}{Line width. Default varies by plot type.}

\item{linetype}{Line type. Default varies by plot type.}

\item{alpha}{Line alpha. Default varies by plot type.}

\item{color}{Line color. Default varies by plot type.}
}
\value{
A \code{pmx_line} element object
}
\description{
Constructor for line layer aesthetics. Used for trend lines, reference lines,
spaghetti lines, and central tendency lines across all plot types.
}
\examples{
pmx_line(linewidth = 1, linetype = "dashed", color = "red")
}
\seealso{
Other element constructors:
\code{\link[=pmx_color]{pmx_color()}},
\code{\link[=pmx_errorbar]{pmx_errorbar()}},
\code{\link[=pmx_point]{pmx_point()}},
\code{\link[=pmx_ribbon]{pmx_ribbon()}},
\code{\link[=pmx_style]{pmx_style()}},
\code{\link[=pmx_trend]{pmx_trend()}}
}
\concept{element constructors}
