% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_element_class.R
\name{pmx_point}
\alias{pmx_point}
\title{Point aesthetics}
\usage{
pmx_point(shape = NULL, size = NULL, alpha = NULL, color = NULL)
}
\arguments{
\item{shape}{Point shape. Default varies by plot type.}

\item{size}{Point size. Default varies by plot type.}

\item{alpha}{Point alpha. Default varies by plot type.}

\item{color}{Point color. Default varies by plot type.}
}
\value{
A \code{pmx_point} element object
}
\description{
Constructor for point layer aesthetics. Used for observed data points and
central tendency points across all plot types.
}
\examples{
pmx_point(shape = 1, size = 2, color = "blue")
}
\seealso{
Other element constructors:
\code{\link[=pmx_color]{pmx_color()}},
\code{\link[=pmx_errorbar]{pmx_errorbar()}},
\code{\link[=pmx_line]{pmx_line()}},
\code{\link[=pmx_ribbon]{pmx_ribbon()}},
\code{\link[=pmx_style]{pmx_style()}},
\code{\link[=pmx_trend]{pmx_trend()}}
}
\concept{element constructors}
