% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxhelpr-package.R
\docType{package}
\name{pmxhelpr-package}
\alias{pmxhelpr}
\alias{pmxhelpr-package}
\title{pmxhelpr: Helper Functions for Pharmacometrics}
\description{
Helper functions to make pharmacometrics workflows more standardized, efficient, and
reproducible. Includes helper functions for exploratory data analysis, goodness-of-fit
diagnostics, ands visual predictive check model evaluation.
}
\section{Getting started}{

Recommended entry points by task:
\itemize{
\item Exploratory PK / PK-PD analysis -- \code{\link[=plot_dvtime]{plot_dvtime()}}, \code{\link[=plot_dvconc]{plot_dvconc()}}
\item Goodness-of-fit diagnostics -- \code{\link[=plot_gof]{plot_gof()}}
\item Visual predictive checks -- \code{\link[=plot_vpc_cont]{plot_vpc_cont()}}, \code{\link[=df_vpcstats]{df_vpcstats()}}
\item Dose-proportionality assessment -- \code{\link[=plot_doseprop]{plot_doseprop()}}, \code{\link[=df_doseprop]{df_doseprop()}}
\item Plot theme customization -- \verb{plot_*_theme()} factories with \verb{pmx_*()}
element constructors
}

Full narrative documentation with worked examples lives on the package
website at \url{https://ryancrass.github.io/pmxhelpr/}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryancrass/pmxhelpr}
  \item \url{https://ryancrass.github.io/pmxhelpr/}
  \item Report bugs at \url{https://github.com/ryancrass/pmxhelpr/issues}
}

}
\author{
\strong{Maintainer}: Ryan Crass \email{ryancrass@gmail.com} (\href{https://orcid.org/0000-0001-5403-4730}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ryan Crass \email{ryancrass@gmail.com} (\href{https://orcid.org/0000-0001-5403-4730}{ORCID}) [copyright holder]
}

}
\keyword{internal}
