% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_addn}
\alias{var_addn}
\title{Append counts of unique identifiers to group labels}
\usage{
var_addn(grp_var, id_var, sep = NULL)
}
\arguments{
\item{grp_var}{Vector of grouping variable values.}

\item{id_var}{Vector of identifier values to count distinct entries of.}

\item{sep}{Optional separator to add between values of \code{grp_var} and appended counts.
Default is NULL.}
}
\value{
A factor vector with group labels appended with subject counts (e.g., \code{"100 mg (n=6)"})
}
\description{
\code{var_addn()} counts distinct values of \code{id_var} within each
level of \code{grp_var} and returns a factor with labels like \code{"100 mg (n=6)"}.
}
\examples{
data <- dplyr::filter(data_sad, CMT != 3)
var_addn(data$DOSE, data$ID, sep = "mg")

}
\seealso{
Other vectorized helpers: 
\code{\link{var_dosenorm}()},
\code{\link{var_predcorr}()}
}
\concept{vectorized helpers}
