% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_predcorr}
\alias{var_predcorr}
\title{Apply prediction correction}
\usage{
var_predcorr(dv_var, pred_var, lower_bound = 0)
}
\arguments{
\item{dv_var}{Vector containing the dependent variable (DV)}

\item{pred_var}{Vector containing population predictions (PRED)}

\item{lower_bound}{Lower bound for prediction correction formula.}
}
\value{
A numeric vector of prediction-corrected values of \code{dv_var}
}
\description{
Apply prediction correction
}
\examples{
pkmodel <- model_mread_load(model = "pkmodel")
data <- df_mrgsim_addpred(data = dplyr::filter(data_sad, CMT != 3), model = pkmodel)
data <- dplyr::mutate(data, PCDV = var_predcorr(ODV, PRED))

}
