% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_plot_databins.R
\name{vpc_plot_databins}
\alias{vpc_plot_databins}
\title{Plot a visual predictive check (VPC) with exact time bins}
\usage{
vpc_plot_databins(
  sim,
  pcvpc = FALSE,
  strat_vars = NULL,
  irep_name = "SIM",
  min_bin_count = 1,
  ...
)
}
\arguments{
\item{sim}{Input simulation dataset. Must contain the following variables: \code{"ID"}, \code{TIME}}

\item{pcvpc}{logical for prediction correction. Default is \code{FALSE}.}

\item{strat_vars}{Character vector of stratification variables passed to \code{stratify} argument
of \code{\link[vpc:vpc]{vpc::vpc()}}. Currently, maximum length = 2.}

\item{irep_name}{Name of replicate variable in \code{sim}. Must be a string. Default is \code{"SIM"}}

\item{min_bin_count}{Minimum number of quantifiable observations in exact bin for inclusion
in binned plot layers. This argument drops small bins from summary statistic calculation
but retains these observations in the observed data points.}

\item{...}{Other arguments passed to \code{\link[vpc:vpc]{vpc::vpc()}}.}
}
\value{
A list containing calculated VPC information (when \code{vpcdb=TRUE}), or a ggplot2 object (default)
}
\description{
\code{vpc_plot_databins()} is a wrapper function for \code{\link[vpc:vpc]{vpc::vpc()}}
that returns a \code{ggplot2} object.
}
\examples{
model <- model_load(model = "model")
simout <- mrgsim_vpc(data = data_sad, model = model, replicates = 100,
output_vars = c(DV = "ODV"),
num_vars = c("CMT", "EVID", "MDV", "NTIME", "LLOQ", "WTBL", "FOOD"),
char_vars = c("USUBJID", "PART"),
irep_name = "SIM")

vpc_plot <- vpc_plot_databins(
sim = simout,
pcvpc = TRUE,
pi = c(0.05, 0.95),
ci = c(0.05, 0.95),
log_y = TRUE)
}
